+++ tool/find-and-replace/__init__.py
@@ -1,0 +1,3 @@
+"""
+Tool: find-and-replace
+"""