+++ tool/rename-files/__init__.py
@@ -1,0 +1,3 @@
+"""
+Tool: rename-files
+"""